refactor: 포스트 카드 컴포넌트 리팩토링
